Changes for token engine in newton and ocata
The default token engine changed to fernet in the newton cycle, so if
the pillar defines some other engine we should respect that. For now
this assumes that other engine is UUID tokens.

The persistence does not apply if using fernet tokens, so don't attempt
to set it.

The memcache_pool token persistence engine was entirely removed in
Ocata; the only engine still available is SQL.

These drivers are loaded through stevedore and issue warnings on ocata
if looking up by classname, so use the entrypoint instead.